[🚚rename]:  pages 하위 폴더 Kit 추가
