+++ src/pages/Kit/KitDetailPage/type.ts
