+++ src/pages/Kit/KitOrderPage/ShippingInfo/DeliveryNote.tsx
