+++ src/pages/index.ts
@@ -1,13 +1,13 @@
 export { default as HomePage } from './HomePage/HomePage';
-export { default as KitDetailPage } from './KitDetailPage/KitDetailPage';
-export { default as KitOrderPage } from './KitOrderPage/KitOrderPage';
+export { default as KitDetailPage } from './Kit/KitDetailPage/KitDetailPage';
+export { default as KitOrderPage } from './Kit/KitOrderPage/KitOrderPage';
+export { default as OrderComplete } from './Kit/OrderComplete/OrderComplete';
 export { default as LoginPage } from './LoginPage/LoginPage';
+export { default as PlantRecommend } from './MyPlantPage/AIRecommend/AIRecommend';
 export { default as MyPlantPage } from './MyPlantPage/MyPlantPage/MyPlantPage';
 export { default as PlantDetail } from './MyPlantPage/PlantDetail/PlantDetail';
-export { default as PlantRecommend } from './MyPlantPage/AIRecommend/AIRecommend';
 export { default as PlantRegister } from './MyPlantPage/PlantRegister/PlantRegister';
 export { default as NotFoundPage } from './NotFoundPage/NotFoundPage';
-export { default as OrderComplete } from './OrderComplete/OrderComplete';
 export { default as OrderList } from './OrderList/OrderList';
 export { default as PasswordChange } from './PasswordChangePage/PasswordChangePage';
 export { default as ProfilePage } from './ProfilePage/ProfilePage';
